requirements.txt
google-ads
psycopg2
python-dotenv